+++ util2/CHANGELOG.md
@@ -36,7 +36,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util2" % "0.5
 | cats-effect |        3.4.10        |               3.4.11 |
 | fs2-io |        3.4.0         |                3.6.1 |
 | sbt-scoverage |        2.0.7         |                2.0.8 |
-| scalatest |        3.2.15        |               3.2.16 |
+| scalatest |        3.2.15        |               3.2.17 |
 
 
 ## 0.3
@@ -67,7 +67,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util2" % "0.3
 | jackson-module-scala |  2.14.1 | 2.15.0 |
 | logstash-logback-encoder |  7.2 | 7.3 |
 | sbt-scoverage |  2.0.6 | 2.0.7 |
-| scalatest |  3.2.14 | 3.2.16 |
+| scalatest |  3.2.14 | 3.2.17 |
 
 ## 0.2
 Breaking Changes: